requirements.txt
google-adk>=0.1.0
google-cloud-bigquery>=3.11.0
google-generativeai>=0.3.0
toolbox-core>=0.1.0
flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
